Minigame Experimentation
Was able to do some experimentation for a possible minigame for both the feeding and drinking. Made two options, but basically the same minigame overall, where the player holds the Space key until it got to the threshold, and another where the player needs to just press the Space key when it got to the threshold. I'm planning the former for the drinking minigame, and the latter being the feeding minigame.